Minor logic update

+++ Scripts/Install/install-send-block-data.sh
@@ -64,7 +64,14 @@ echo "***"
 echo -e "${NC}"
 _sleep 2
 sudo test -d /mnt/usb1/system-setup-salvage || sudo mkdir /mnt/usb1/system-setup-salvage
-sudo rsync -vahW --no-compress --progress --delete-after /mnt/usb/docker/volumes/my-dojo_data-bitcoind/_data/{blocks,chainstate} /mnt/usb1/system-setup-salvage
+
+if sudo test -d /mnt/usb1/system-setup-salvage/blocks; then
+    # Use rsync when existing IBD is found
+    sudo rsync -vahW --no-compress --progress --delete-after /mnt/usb/docker/volumes/my-dojo_data-bitcoind/_data/{blocks,chainstate} /mnt/usb1/system-setup-salvage
+else
+    # Use cp for initial fresh IBD copy
+    sudo cp -av /mnt/usb/docker/volumes/my-dojo_data-bitcoind/_data/{blocks,chainstate} /mnt/usb1/system-setup-salvage
+fi
 # copies blockchain data to backup drive while keeping permissions so we can later restore properly
 
 echo -e "${RED}"
